Overhaul structure of renderers
