+++ data/skeletons/BASE.csv
@@ -1,0 +1,1 @@
+name,parent,swap
