+++ data/skeletons/BASE.json
@@ -1,0 +1,66 @@
+{
+    "markers": { "height": 0.005, "radius": 0.005 },
+    "keypoints": {},
+    "joints": {
+        "S": {
+            "type": 2,
+            "max": 2,
+            "min": -2,
+            "parent": null,
+            "parent_mult": 0,
+            "parent_offset": 0,
+            "self_mult": 1,
+            "predictors": {}
+        },
+        "L": {
+            "type": 1,
+            "max": 4,
+            "min": -4,
+            "parent": null,
+            "parent_mult": 0,
+            "parent_offset": 0,
+            "self_mult": 1,
+            "predictors": {}
+        },
+        "U": {
+            "type": 1,
+            "max": 4,
+            "min": -4,
+            "parent": "L",
+            "parent_mult": 1,
+            "parent_offset": 0,
+            "self_mult": 1,
+            "predictors": {}
+        },
+        "R": {
+            "type": 3,
+            "max": 2,
+            "min": -2,
+            "parent": null,
+            "parent_mult": 0,
+            "parent_offset": 0,
+            "self_mult": 1,
+            "predictors": {}
+        },
+        "B": {
+            "type": 1,
+            "max": 4,
+            "min": -4,
+            "parent": "U",
+            "parent_mult": 1,
+            "parent_offset": 0,
+            "self_mult": 1,
+            "predictors": {}
+        },
+        "T": {
+            "type": 3,
+            "max": 2,
+            "min": -2,
+            "parent": null,
+            "parent_mult": 0,
+            "parent_offset": 0,
+            "self_mult": 1,
+            "predictors": {}
+        }
+    }
+}
