+++ robotpose/segmentation.py
@@ -13,12 +13,8 @@ from pixellib.instance import custom_segmentation
 import cv2
 import os
 import numpy as np
-import open3d as o3d
 import pyrealsense2 as rs
-from .utils import Timer, vizDepth_new
-from tqdm import tqdm
 from . import paths as p
-import time
 from . import projection as proj
